+++ indico/modules/core/client/js/index.js
@@ -7,4 +7,5 @@
 
 import './impersonation';
 import './session_bar';
+import './session_expiry';
 import './top_bars';