+++ indico/web/flask/session.py
@@ -7,6 +7,7 @@
 
 import functools
 import pickle
+import re
 import uuid
 from datetime import datetime, timedelta, timezone
 from pathlib import Path
@@ -20,9 +21,12 @@ from werkzeug.utils import cached_property
 from indico.core.cache import make_scoped_cache
 from indico.core.config import config
 from indico.modules.users import User
-from indico.util.date_time import get_display_tz
+from indico.util.date_time import get_display_tz, utc_to_server
 from indico.util.i18n import set_best_lang
 from indico.web.util import get_request_user
+
+
+RE_SKIP_REFRESH_SESSION_FOR_ASSETS_ENDPOINTS = re.compile(r'assets\.|plugin_.*\.static$')
 
 
 @functools.cache
@@ -182,13 +186,23 @@ class IndicoSessionInterface(SessionInterface):
             session_lifetime = self.temporary_session_lifetime
 
         if session.hard_expiry:
-            return min(session_lifetime, session.hard_expiry - datetime.now(timezone.utc))
+            hard_lifetime = session.hard_expiry - datetime.now(timezone.utc)
+            if not session_lifetime:
+                # if we have `SESSION_LIFETIME = 0` ("browser session"), the `min()` logic below
+                # would not work properly because 0 generally means 'no expiry'
+                return hard_lifetime
+            return min(session_lifetime, hard_lifetime)
 
         return session_lifetime
 
     def should_refresh_session(self, app, session):
         if session.new or '_expires' not in session:
             return False
+        if (request.endpoint == 'core.session_expiry' or
+                RE_SKIP_REFRESH_SESSION_FOR_ASSETS_ENDPOINTS.match(request.endpoint)):
+            return False
+        if request.endpoint == 'core.session_refresh':
+            return True
         threshold = self.get_storage_lifetime(app, session) / 2
         return session['_expires'] - datetime.now() < threshold
 
@@ -241,6 +255,7 @@ class IndicoSessionInterface(SessionInterface):
         storage_ttl = self.get_storage_lifetime(app, session)
         if session.hard_expiry:
             cookie_lifetime = session.hard_expiry
+            session['_expires'] = utc_to_server(session.hard_expiry).replace(tzinfo=None)
         else:
             cookie_lifetime = self.get_expiration_time(app, session)
             session['_expires'] = datetime.now() + storage_ttl