+++ CHANGES.rst
@@ -99,6 +99,8 @@ Improvements
 - Allow linking badge templates to registration forms in order to use custom fields in them
   (:pr:`6088`)
 - Allow filtering the list of editables by tags (:issue:`6195`, :pr:`6197`)
+- Warn users with a dialog before their session expires and let them extend it (:pr:`6026`,
+  thanks :user:`SegiNyn`)
 
 Bugfixes
 ^^^^^^^^